added an args val to find-bearer-token and a new key to defapi
the new argument to find-bearer-token exists to allow the user
to specialize their bearer token finding mechanism on a per
api request. Perhaps on one api they want to search in a different
table for a bearer token, to do this they can make the key
bearer-verifier-args something like '(:admin) and this will be
passed as that list to find-bearer-token where they could have a
method that specializes on a keyword.